+++ lib/model/om/BaseClassStudent.php
@@ -538,6 +538,11 @@ abstract class BaseClassStudent extends BaseObject  implements Persistent
 			$deleteQuery = ClassStudentQuery::create()
 				->filterByPrimaryKey($this->getPrimaryKey());
 			$ret = $this->preDelete($con);
+			// denyable behavior
+			if(!(ClassStudentPeer::isIgnoringRights() || $this->mayOperate("delete"))) {
+				throw new PropelException(new NotPermittedException("delete.by_role", array("role_key" => "class_students")));
+			}
+
 			if ($ret) {
 				$deleteQuery->delete($con);
 				$this->postDelete($con);
@@ -581,6 +586,11 @@ abstract class BaseClassStudent extends BaseObject  implements Persistent
 			$ret = $this->preSave($con);
 			if ($isInsert) {
 				$ret = $ret && $this->preInsert($con);
+				// denyable behavior
+				if(!(ClassStudentPeer::isIgnoringRights() || $this->mayOperate("insert"))) {
+					throw new PropelException(new NotPermittedException("insert.by_role", array("role_key" => "class_students")));
+				}
+
 				// extended_timestampable behavior
 				if (!$this->isColumnModified(ClassStudentPeer::CREATED_AT)) {
 					$this->setCreatedAt(time());
@@ -601,6 +611,11 @@ abstract class BaseClassStudent extends BaseObject  implements Persistent
 
 			} else {
 				$ret = $ret && $this->preUpdate($con);
+				// denyable behavior
+				if(!(ClassStudentPeer::isIgnoringRights() || $this->mayOperate("update"))) {
+					throw new PropelException(new NotPermittedException("update.by_role", array("role_key" => "class_students")));
+				}
+
 				// extended_timestampable behavior
 				if ($this->isModified() && !$this->isColumnModified(ClassStudentPeer::UPDATED_AT)) {
 					$this->setUpdatedAt(time());
@@ -1367,6 +1382,26 @@ abstract class BaseClassStudent extends BaseObject  implements Persistent
 	public function __toString()
 	{
 		return (string) $this->exportTo(ClassStudentPeer::DEFAULT_STRING_FORMAT);
+	}
+
+	// denyable behavior
+	public function mayOperate($sOperation, $oUser = false) {
+		if($oUser === false) {
+			$oUser = Session::getSession()->getUser();
+		}
+		if($oUser && ($this->isNew() || $this->getCreatedBy() === $oUser->getId()) && ClassStudentPeer::mayOperateOnOwn($oUser, $this, $sOperation)) {
+			return true;
+		}
+		return ClassStudentPeer::mayOperateOn($oUser, $this, $sOperation);
+	}
+	public function mayBeInserted($oUser = false) {
+		return $this->mayOperate($oUser, "insert");
+	}
+	public function mayBeUpdated($oUser = false) {
+		return $this->mayOperate($oUser, "update");
+	}
+	public function mayBeDeleted($oUser = false) {
+		return $this->mayOperate($oUser, "delete");
 	}
 
 	// extended_timestampable behavior
