+++ lib/model/om/BaseClassTeacher.php
@@ -646,6 +646,11 @@ abstract class BaseClassTeacher extends BaseObject  implements Persistent
 			$deleteQuery = ClassTeacherQuery::create()
 				->filterByPrimaryKey($this->getPrimaryKey());
 			$ret = $this->preDelete($con);
+			// denyable behavior
+			if(!(ClassTeacherPeer::isIgnoringRights() || $this->mayOperate("delete"))) {
+				throw new PropelException(new NotPermittedException("delete.by_role", array("role_key" => "class_teachers")));
+			}
+
 			if ($ret) {
 				$deleteQuery->delete($con);
 				$this->postDelete($con);
@@ -689,6 +694,11 @@ abstract class BaseClassTeacher extends BaseObject  implements Persistent
 			$ret = $this->preSave($con);
 			if ($isInsert) {
 				$ret = $ret && $this->preInsert($con);
+				// denyable behavior
+				if(!(ClassTeacherPeer::isIgnoringRights() || $this->mayOperate("insert"))) {
+					throw new PropelException(new NotPermittedException("insert.by_role", array("role_key" => "class_teachers")));
+				}
+
 				// extended_timestampable behavior
 				if (!$this->isColumnModified(ClassTeacherPeer::CREATED_AT)) {
 					$this->setCreatedAt(time());
@@ -709,6 +719,11 @@ abstract class BaseClassTeacher extends BaseObject  implements Persistent
 
 			} else {
 				$ret = $ret && $this->preUpdate($con);
+				// denyable behavior
+				if(!(ClassTeacherPeer::isIgnoringRights() || $this->mayOperate("update"))) {
+					throw new PropelException(new NotPermittedException("update.by_role", array("role_key" => "class_teachers")));
+				}
+
 				// extended_timestampable behavior
 				if ($this->isModified() && !$this->isColumnModified(ClassTeacherPeer::UPDATED_AT)) {
 					$this->setUpdatedAt(time());
@@ -1501,6 +1516,26 @@ abstract class BaseClassTeacher extends BaseObject  implements Persistent
 	public function __toString()
 	{
 		return (string) $this->exportTo(ClassTeacherPeer::DEFAULT_STRING_FORMAT);
+	}
+
+	// denyable behavior
+	public function mayOperate($sOperation, $oUser = false) {
+		if($oUser === false) {
+			$oUser = Session::getSession()->getUser();
+		}
+		if($oUser && ($this->isNew() || $this->getCreatedBy() === $oUser->getId()) && ClassTeacherPeer::mayOperateOnOwn($oUser, $this, $sOperation)) {
+			return true;
+		}
+		return ClassTeacherPeer::mayOperateOn($oUser, $this, $sOperation);
+	}
+	public function mayBeInserted($oUser = false) {
+		return $this->mayOperate($oUser, "insert");
+	}
+	public function mayBeUpdated($oUser = false) {
+		return $this->mayOperate($oUser, "update");
+	}
+	public function mayBeDeleted($oUser = false) {
+		return $this->mayOperate($oUser, "delete");
 	}
 
 	// extended_timestampable behavior
