+++ lib/model/om/BaseClassType.php
@@ -556,6 +556,11 @@ abstract class BaseClassType extends BaseObject  implements Persistent
 			$deleteQuery = ClassTypeQuery::create()
 				->filterByPrimaryKey($this->getPrimaryKey());
 			$ret = $this->preDelete($con);
+			// denyable behavior
+			if(!(ClassTypePeer::isIgnoringRights() || $this->mayOperate("delete"))) {
+				throw new PropelException(new NotPermittedException("delete.by_role", array("role_key" => "class_types")));
+			}
+
 			if ($ret) {
 				$deleteQuery->delete($con);
 				$this->postDelete($con);
@@ -599,6 +604,11 @@ abstract class BaseClassType extends BaseObject  implements Persistent
 			$ret = $this->preSave($con);
 			if ($isInsert) {
 				$ret = $ret && $this->preInsert($con);
+				// denyable behavior
+				if(!(ClassTypePeer::isIgnoringRights() || $this->mayOperate("insert"))) {
+					throw new PropelException(new NotPermittedException("insert.by_role", array("role_key" => "class_types")));
+				}
+
 				// extended_timestampable behavior
 				if (!$this->isColumnModified(ClassTypePeer::CREATED_AT)) {
 					$this->setCreatedAt(time());
@@ -619,6 +629,11 @@ abstract class BaseClassType extends BaseObject  implements Persistent
 
 			} else {
 				$ret = $ret && $this->preUpdate($con);
+				// denyable behavior
+				if(!(ClassTypePeer::isIgnoringRights() || $this->mayOperate("update"))) {
+					throw new PropelException(new NotPermittedException("update.by_role", array("role_key" => "class_types")));
+				}
+
 				// extended_timestampable behavior
 				if ($this->isModified() && !$this->isColumnModified(ClassTypePeer::UPDATED_AT)) {
 					$this->setUpdatedAt(time());
@@ -1614,6 +1629,26 @@ abstract class BaseClassType extends BaseObject  implements Persistent
 	public function __toString()
 	{
 		return (string) $this->exportTo(ClassTypePeer::DEFAULT_STRING_FORMAT);
+	}
+
+	// denyable behavior
+	public function mayOperate($sOperation, $oUser = false) {
+		if($oUser === false) {
+			$oUser = Session::getSession()->getUser();
+		}
+		if($oUser && ($this->isNew() || $this->getCreatedBy() === $oUser->getId()) && ClassTypePeer::mayOperateOnOwn($oUser, $this, $sOperation)) {
+			return true;
+		}
+		return ClassTypePeer::mayOperateOn($oUser, $this, $sOperation);
+	}
+	public function mayBeInserted($oUser = false) {
+		return $this->mayOperate($oUser, "insert");
+	}
+	public function mayBeUpdated($oUser = false) {
+		return $this->mayOperate($oUser, "update");
+	}
+	public function mayBeDeleted($oUser = false) {
+		return $this->mayOperate($oUser, "delete");
 	}
 
 	// extended_timestampable behavior
