+++ lib/model/om/BaseEvent.php
@@ -1178,6 +1178,11 @@ abstract class BaseEvent extends BaseObject  implements Persistent
 			$deleteQuery = EventQuery::create()
 				->filterByPrimaryKey($this->getPrimaryKey());
 			$ret = $this->preDelete($con);
+			// denyable behavior
+			if(!(EventPeer::isIgnoringRights() || $this->mayOperate("delete"))) {
+				throw new PropelException(new NotPermittedException("delete.by_role", array("role_key" => "events")));
+			}
+
 			if ($ret) {
 				$deleteQuery->delete($con);
 				$this->postDelete($con);
@@ -1221,6 +1226,11 @@ abstract class BaseEvent extends BaseObject  implements Persistent
 			$ret = $this->preSave($con);
 			if ($isInsert) {
 				$ret = $ret && $this->preInsert($con);
+				// denyable behavior
+				if(!(EventPeer::isIgnoringRights() || $this->mayOperate("insert"))) {
+					throw new PropelException(new NotPermittedException("insert.by_role", array("role_key" => "events")));
+				}
+
 				// extended_timestampable behavior
 				if (!$this->isColumnModified(EventPeer::CREATED_AT)) {
 					$this->setCreatedAt(time());
@@ -1241,6 +1251,11 @@ abstract class BaseEvent extends BaseObject  implements Persistent
 
 			} else {
 				$ret = $ret && $this->preUpdate($con);
+				// denyable behavior
+				if(!(EventPeer::isIgnoringRights() || $this->mayOperate("update"))) {
+					throw new PropelException(new NotPermittedException("update.by_role", array("role_key" => "events")));
+				}
+
 				// extended_timestampable behavior
 				if ($this->isModified() && !$this->isColumnModified(EventPeer::UPDATED_AT)) {
 					$this->setUpdatedAt(time());
@@ -2543,6 +2558,26 @@ abstract class BaseEvent extends BaseObject  implements Persistent
 	public function __toString()
 	{
 		return (string) $this->exportTo(EventPeer::DEFAULT_STRING_FORMAT);
+	}
+
+	// denyable behavior
+	public function mayOperate($sOperation, $oUser = false) {
+		if($oUser === false) {
+			$oUser = Session::getSession()->getUser();
+		}
+		if($oUser && ($this->isNew() || $this->getCreatedBy() === $oUser->getId()) && EventPeer::mayOperateOnOwn($oUser, $this, $sOperation)) {
+			return true;
+		}
+		return EventPeer::mayOperateOn($oUser, $this, $sOperation);
+	}
+	public function mayBeInserted($oUser = false) {
+		return $this->mayOperate($oUser, "insert");
+	}
+	public function mayBeUpdated($oUser = false) {
+		return $this->mayOperate($oUser, "update");
+	}
+	public function mayBeDeleted($oUser = false) {
+		return $this->mayOperate($oUser, "delete");
 	}
 
 	// extended_timestampable behavior
