+++ lib/model/om/BaseNote.php
@@ -618,6 +618,11 @@ abstract class BaseNote extends BaseObject  implements Persistent
 			$deleteQuery = NoteQuery::create()
 				->filterByPrimaryKey($this->getPrimaryKey());
 			$ret = $this->preDelete($con);
+			// denyable behavior
+			if(!(NotePeer::isIgnoringRights() || $this->mayOperate("delete"))) {
+				throw new PropelException(new NotPermittedException("delete.by_role", array("role_key" => "notes")));
+			}
+
 			if ($ret) {
 				$deleteQuery->delete($con);
 				$this->postDelete($con);
@@ -661,6 +666,11 @@ abstract class BaseNote extends BaseObject  implements Persistent
 			$ret = $this->preSave($con);
 			if ($isInsert) {
 				$ret = $ret && $this->preInsert($con);
+				// denyable behavior
+				if(!(NotePeer::isIgnoringRights() || $this->mayOperate("insert"))) {
+					throw new PropelException(new NotPermittedException("insert.by_role", array("role_key" => "notes")));
+				}
+
 				// extended_timestampable behavior
 				if (!$this->isColumnModified(NotePeer::CREATED_AT)) {
 					$this->setCreatedAt(time());
@@ -681,6 +691,11 @@ abstract class BaseNote extends BaseObject  implements Persistent
 
 			} else {
 				$ret = $ret && $this->preUpdate($con);
+				// denyable behavior
+				if(!(NotePeer::isIgnoringRights() || $this->mayOperate("update"))) {
+					throw new PropelException(new NotPermittedException("update.by_role", array("role_key" => "notes")));
+				}
+
 				// extended_timestampable behavior
 				if ($this->isModified() && !$this->isColumnModified(NotePeer::UPDATED_AT)) {
 					$this->setUpdatedAt(time());
@@ -1332,6 +1347,26 @@ abstract class BaseNote extends BaseObject  implements Persistent
 	public function __toString()
 	{
 		return (string) $this->exportTo(NotePeer::DEFAULT_STRING_FORMAT);
+	}
+
+	// denyable behavior
+	public function mayOperate($sOperation, $oUser = false) {
+		if($oUser === false) {
+			$oUser = Session::getSession()->getUser();
+		}
+		if($oUser && ($this->isNew() || $this->getCreatedBy() === $oUser->getId()) && NotePeer::mayOperateOnOwn($oUser, $this, $sOperation)) {
+			return true;
+		}
+		return NotePeer::mayOperateOn($oUser, $this, $sOperation);
+	}
+	public function mayBeInserted($oUser = false) {
+		return $this->mayOperate($oUser, "insert");
+	}
+	public function mayBeUpdated($oUser = false) {
+		return $this->mayOperate($oUser, "update");
+	}
+	public function mayBeDeleted($oUser = false) {
+		return $this->mayOperate($oUser, "delete");
 	}
 
 	// extended_timestampable behavior
