+++ lib/model/om/BaseSchool.php
@@ -785,6 +785,11 @@ abstract class BaseSchool extends BaseObject  implements Persistent
 			$deleteQuery = SchoolQuery::create()
 				->filterByPrimaryKey($this->getPrimaryKey());
 			$ret = $this->preDelete($con);
+			// denyable behavior
+			if(!(SchoolPeer::isIgnoringRights() || $this->mayOperate("delete"))) {
+				throw new PropelException(new NotPermittedException("delete.by_role", array("role_key" => "schools")));
+			}
+
 			if ($ret) {
 				$deleteQuery->delete($con);
 				$this->postDelete($con);
@@ -828,6 +833,11 @@ abstract class BaseSchool extends BaseObject  implements Persistent
 			$ret = $this->preSave($con);
 			if ($isInsert) {
 				$ret = $ret && $this->preInsert($con);
+				// denyable behavior
+				if(!(SchoolPeer::isIgnoringRights() || $this->mayOperate("insert"))) {
+					throw new PropelException(new NotPermittedException("insert.by_role", array("role_key" => "schools")));
+				}
+
 				// extended_timestampable behavior
 				if (!$this->isColumnModified(SchoolPeer::CREATED_AT)) {
 					$this->setCreatedAt(time());
@@ -848,6 +858,11 @@ abstract class BaseSchool extends BaseObject  implements Persistent
 
 			} else {
 				$ret = $ret && $this->preUpdate($con);
+				// denyable behavior
+				if(!(SchoolPeer::isIgnoringRights() || $this->mayOperate("update"))) {
+					throw new PropelException(new NotPermittedException("update.by_role", array("role_key" => "schools")));
+				}
+
 				// extended_timestampable behavior
 				if ($this->isModified() && !$this->isColumnModified(SchoolPeer::UPDATED_AT)) {
 					$this->setUpdatedAt(time());
@@ -2137,6 +2152,26 @@ abstract class BaseSchool extends BaseObject  implements Persistent
 	public function __toString()
 	{
 		return (string) $this->exportTo(SchoolPeer::DEFAULT_STRING_FORMAT);
+	}
+
+	// denyable behavior
+	public function mayOperate($sOperation, $oUser = false) {
+		if($oUser === false) {
+			$oUser = Session::getSession()->getUser();
+		}
+		if($oUser && ($this->isNew() || $this->getCreatedBy() === $oUser->getId()) && SchoolPeer::mayOperateOnOwn($oUser, $this, $sOperation)) {
+			return true;
+		}
+		return SchoolPeer::mayOperateOn($oUser, $this, $sOperation);
+	}
+	public function mayBeInserted($oUser = false) {
+		return $this->mayOperate($oUser, "insert");
+	}
+	public function mayBeUpdated($oUser = false) {
+		return $this->mayOperate($oUser, "update");
+	}
+	public function mayBeDeleted($oUser = false) {
+		return $this->mayOperate($oUser, "delete");
 	}
 
 	// extended_timestampable behavior
