+++ lib/model/om/BaseSchoolBuilding.php
@@ -519,6 +519,11 @@ abstract class BaseSchoolBuilding extends BaseObject  implements Persistent
 			$deleteQuery = SchoolBuildingQuery::create()
 				->filterByPrimaryKey($this->getPrimaryKey());
 			$ret = $this->preDelete($con);
+			// denyable behavior
+			if(!(SchoolBuildingPeer::isIgnoringRights() || $this->mayOperate("delete"))) {
+				throw new PropelException(new NotPermittedException("delete.by_role", array("role_key" => "school_buildings")));
+			}
+
 			if ($ret) {
 				$deleteQuery->delete($con);
 				$this->postDelete($con);
@@ -562,6 +567,11 @@ abstract class BaseSchoolBuilding extends BaseObject  implements Persistent
 			$ret = $this->preSave($con);
 			if ($isInsert) {
 				$ret = $ret && $this->preInsert($con);
+				// denyable behavior
+				if(!(SchoolBuildingPeer::isIgnoringRights() || $this->mayOperate("insert"))) {
+					throw new PropelException(new NotPermittedException("insert.by_role", array("role_key" => "school_buildings")));
+				}
+
 				// extended_timestampable behavior
 				if (!$this->isColumnModified(SchoolBuildingPeer::CREATED_AT)) {
 					$this->setCreatedAt(time());
@@ -582,6 +592,11 @@ abstract class BaseSchoolBuilding extends BaseObject  implements Persistent
 
 			} else {
 				$ret = $ret && $this->preUpdate($con);
+				// denyable behavior
+				if(!(SchoolBuildingPeer::isIgnoringRights() || $this->mayOperate("update"))) {
+					throw new PropelException(new NotPermittedException("update.by_role", array("role_key" => "school_buildings")));
+				}
+
 				// extended_timestampable behavior
 				if ($this->isModified() && !$this->isColumnModified(SchoolBuildingPeer::UPDATED_AT)) {
 					$this->setUpdatedAt(time());
@@ -1566,6 +1581,26 @@ abstract class BaseSchoolBuilding extends BaseObject  implements Persistent
 	public function __toString()
 	{
 		return (string) $this->exportTo(SchoolBuildingPeer::DEFAULT_STRING_FORMAT);
+	}
+
+	// denyable behavior
+	public function mayOperate($sOperation, $oUser = false) {
+		if($oUser === false) {
+			$oUser = Session::getSession()->getUser();
+		}
+		if($oUser && ($this->isNew() || $this->getCreatedBy() === $oUser->getId()) && SchoolBuildingPeer::mayOperateOnOwn($oUser, $this, $sOperation)) {
+			return true;
+		}
+		return SchoolBuildingPeer::mayOperateOn($oUser, $this, $sOperation);
+	}
+	public function mayBeInserted($oUser = false) {
+		return $this->mayOperate($oUser, "insert");
+	}
+	public function mayBeUpdated($oUser = false) {
+		return $this->mayOperate($oUser, "update");
+	}
+	public function mayBeDeleted($oUser = false) {
+		return $this->mayOperate($oUser, "delete");
 	}
 
 	// extended_timestampable behavior
