+++ lib/model/om/BaseSchoolClass.php
@@ -1062,6 +1062,11 @@ abstract class BaseSchoolClass extends BaseObject  implements Persistent
 			$deleteQuery = SchoolClassQuery::create()
 				->filterByPrimaryKey($this->getPrimaryKey());
 			$ret = $this->preDelete($con);
+			// denyable behavior
+			if(!(SchoolClassPeer::isIgnoringRights() || $this->mayOperate("delete"))) {
+				throw new PropelException(new NotPermittedException("delete.by_role", array("role_key" => "school_classes")));
+			}
+
 			if ($ret) {
 				$deleteQuery->delete($con);
 				$this->postDelete($con);
@@ -1105,6 +1110,11 @@ abstract class BaseSchoolClass extends BaseObject  implements Persistent
 			$ret = $this->preSave($con);
 			if ($isInsert) {
 				$ret = $ret && $this->preInsert($con);
+				// denyable behavior
+				if(!(SchoolClassPeer::isIgnoringRights() || $this->mayOperate("insert"))) {
+					throw new PropelException(new NotPermittedException("insert.by_role", array("role_key" => "school_classes")));
+				}
+
 				// extended_timestampable behavior
 				if (!$this->isColumnModified(SchoolClassPeer::CREATED_AT)) {
 					$this->setCreatedAt(time());
@@ -1125,6 +1135,11 @@ abstract class BaseSchoolClass extends BaseObject  implements Persistent
 
 			} else {
 				$ret = $ret && $this->preUpdate($con);
+				// denyable behavior
+				if(!(SchoolClassPeer::isIgnoringRights() || $this->mayOperate("update"))) {
+					throw new PropelException(new NotPermittedException("update.by_role", array("role_key" => "school_classes")));
+				}
+
 				// extended_timestampable behavior
 				if ($this->isModified() && !$this->isColumnModified(SchoolClassPeer::UPDATED_AT)) {
 					$this->setUpdatedAt(time());
@@ -3271,6 +3286,26 @@ abstract class BaseSchoolClass extends BaseObject  implements Persistent
 	public function __toString()
 	{
 		return (string) $this->exportTo(SchoolClassPeer::DEFAULT_STRING_FORMAT);
+	}
+
+	// denyable behavior
+	public function mayOperate($sOperation, $oUser = false) {
+		if($oUser === false) {
+			$oUser = Session::getSession()->getUser();
+		}
+		if($oUser && ($this->isNew() || $this->getCreatedBy() === $oUser->getId()) && SchoolClassPeer::mayOperateOnOwn($oUser, $this, $sOperation)) {
+			return true;
+		}
+		return SchoolClassPeer::mayOperateOn($oUser, $this, $sOperation);
+	}
+	public function mayBeInserted($oUser = false) {
+		return $this->mayOperate($oUser, "insert");
+	}
+	public function mayBeUpdated($oUser = false) {
+		return $this->mayOperate($oUser, "update");
+	}
+	public function mayBeDeleted($oUser = false) {
+		return $this->mayOperate($oUser, "delete");
 	}
 
 	// extended_timestampable behavior
