+++ lib/model/om/BaseSchoolFunction.php
@@ -619,6 +619,11 @@ abstract class BaseSchoolFunction extends BaseObject  implements Persistent
 			$deleteQuery = SchoolFunctionQuery::create()
 				->filterByPrimaryKey($this->getPrimaryKey());
 			$ret = $this->preDelete($con);
+			// denyable behavior
+			if(!(SchoolFunctionPeer::isIgnoringRights() || $this->mayOperate("delete"))) {
+				throw new PropelException(new NotPermittedException("delete.by_role", array("role_key" => "school_functions")));
+			}
+
 			if ($ret) {
 				$deleteQuery->delete($con);
 				$this->postDelete($con);
@@ -662,6 +667,11 @@ abstract class BaseSchoolFunction extends BaseObject  implements Persistent
 			$ret = $this->preSave($con);
 			if ($isInsert) {
 				$ret = $ret && $this->preInsert($con);
+				// denyable behavior
+				if(!(SchoolFunctionPeer::isIgnoringRights() || $this->mayOperate("insert"))) {
+					throw new PropelException(new NotPermittedException("insert.by_role", array("role_key" => "school_functions")));
+				}
+
 				// extended_timestampable behavior
 				if (!$this->isColumnModified(SchoolFunctionPeer::CREATED_AT)) {
 					$this->setCreatedAt(time());
@@ -682,6 +692,11 @@ abstract class BaseSchoolFunction extends BaseObject  implements Persistent
 
 			} else {
 				$ret = $ret && $this->preUpdate($con);
+				// denyable behavior
+				if(!(SchoolFunctionPeer::isIgnoringRights() || $this->mayOperate("update"))) {
+					throw new PropelException(new NotPermittedException("update.by_role", array("role_key" => "school_functions")));
+				}
+
 				// extended_timestampable behavior
 				if ($this->isModified() && !$this->isColumnModified(SchoolFunctionPeer::UPDATED_AT)) {
 					$this->setUpdatedAt(time());
@@ -1720,6 +1735,26 @@ abstract class BaseSchoolFunction extends BaseObject  implements Persistent
 	public function __toString()
 	{
 		return (string) $this->exportTo(SchoolFunctionPeer::DEFAULT_STRING_FORMAT);
+	}
+
+	// denyable behavior
+	public function mayOperate($sOperation, $oUser = false) {
+		if($oUser === false) {
+			$oUser = Session::getSession()->getUser();
+		}
+		if($oUser && ($this->isNew() || $this->getCreatedBy() === $oUser->getId()) && SchoolFunctionPeer::mayOperateOnOwn($oUser, $this, $sOperation)) {
+			return true;
+		}
+		return SchoolFunctionPeer::mayOperateOn($oUser, $this, $sOperation);
+	}
+	public function mayBeInserted($oUser = false) {
+		return $this->mayOperate($oUser, "insert");
+	}
+	public function mayBeUpdated($oUser = false) {
+		return $this->mayOperate($oUser, "update");
+	}
+	public function mayBeDeleted($oUser = false) {
+		return $this->mayOperate($oUser, "delete");
 	}
 
 	// extended_timestampable behavior
