+++ lib/model/om/BaseService.php
@@ -972,6 +972,11 @@ abstract class BaseService extends BaseObject  implements Persistent
 			$deleteQuery = ServiceQuery::create()
 				->filterByPrimaryKey($this->getPrimaryKey());
 			$ret = $this->preDelete($con);
+			// denyable behavior
+			if(!(ServicePeer::isIgnoringRights() || $this->mayOperate("delete"))) {
+				throw new PropelException(new NotPermittedException("delete.by_role", array("role_key" => "services")));
+			}
+
 			if ($ret) {
 				$deleteQuery->delete($con);
 				$this->postDelete($con);
@@ -1015,6 +1020,11 @@ abstract class BaseService extends BaseObject  implements Persistent
 			$ret = $this->preSave($con);
 			if ($isInsert) {
 				$ret = $ret && $this->preInsert($con);
+				// denyable behavior
+				if(!(ServicePeer::isIgnoringRights() || $this->mayOperate("insert"))) {
+					throw new PropelException(new NotPermittedException("insert.by_role", array("role_key" => "services")));
+				}
+
 				// extended_timestampable behavior
 				if (!$this->isColumnModified(ServicePeer::CREATED_AT)) {
 					$this->setCreatedAt(time());
@@ -1035,6 +1045,11 @@ abstract class BaseService extends BaseObject  implements Persistent
 
 			} else {
 				$ret = $ret && $this->preUpdate($con);
+				// denyable behavior
+				if(!(ServicePeer::isIgnoringRights() || $this->mayOperate("update"))) {
+					throw new PropelException(new NotPermittedException("update.by_role", array("role_key" => "services")));
+				}
+
 				// extended_timestampable behavior
 				if ($this->isModified() && !$this->isColumnModified(ServicePeer::UPDATED_AT)) {
 					$this->setUpdatedAt(time());
@@ -2673,6 +2688,26 @@ abstract class BaseService extends BaseObject  implements Persistent
 	public function __toString()
 	{
 		return (string) $this->exportTo(ServicePeer::DEFAULT_STRING_FORMAT);
+	}
+
+	// denyable behavior
+	public function mayOperate($sOperation, $oUser = false) {
+		if($oUser === false) {
+			$oUser = Session::getSession()->getUser();
+		}
+		if($oUser && ($this->isNew() || $this->getCreatedBy() === $oUser->getId()) && ServicePeer::mayOperateOnOwn($oUser, $this, $sOperation)) {
+			return true;
+		}
+		return ServicePeer::mayOperateOn($oUser, $this, $sOperation);
+	}
+	public function mayBeInserted($oUser = false) {
+		return $this->mayOperate($oUser, "insert");
+	}
+	public function mayBeUpdated($oUser = false) {
+		return $this->mayOperate($oUser, "update");
+	}
+	public function mayBeDeleted($oUser = false) {
+		return $this->mayOperate($oUser, "delete");
 	}
 
 	// extended_timestampable behavior
